refactor: move gson from syncwrap to root.
