+++ gmap/gmap_example_test.go
@@ -19,7 +19,7 @@ import (
 	"strconv"
 
 	"github.com/bytedance/gg/gslice"
-	"github.com/bytedance/gg/stdwrap/gson"
+	"github.com/bytedance/gg/gson"
 )
 
 func Example() {
